moving publish function
