+++ example/app/config.yml
@@ -157,8 +157,6 @@ kk-uat-deployment:
     stepRoleArn: 'arn:aws:iam::{{AWS_ACCOUNT_ID}}:role/kk-test-uat-steprole'
     instanceProfile: 'arn:aws:iam::{{AWS_ACCOUNT_ID}}:instance-profile/kk-test-uat-ecs'
     distributionRoleArn: 'arn:aws:iam::{{AWS_ACCOUNT_ID}}:role/kk-test-uat-distribution-api-lambda'
-
-  es: null
    
 mth-cum-test:
   prefix: mth-cum-test
